Fixing the bug #3
With this commit i am trying to fix the bug #3. The bug
causes the frontend to die with an exception if the user
does the database update not fast enough.

Also i am changing the linebreaks to to the default PHP
constant. Now windows servers also return the right
linebreak.

Also i am fixng some typos in the the phpdoc comments.

Signed-off-by: Leo Unglaub <leo.unglaub@2create.at>